backend/requirements.txt
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
docopt==0.6.2
fiona==1.9.5
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.14.2
geopandas==0.14.1
greenlet==3.0.2
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pipreqs==0.4.13
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
